Fix INT → BIGINT mismatch in trx_result type and endpoint
Align trx_result.id (BIGSERIAL), trx_result.subsequent_no (BIGINT),
and get_trx_results last_id parameter to match backing table types.

+++ endpoints/get_trx_results.sql
@@ -68,7 +68,7 @@ DROP FUNCTION IF EXISTS nfttracker_endpoints.get_trx_results;
 CREATE OR REPLACE FUNCTION nfttracker_endpoints.get_trx_results(
     "trx_id" TEXT,
     "count" INT = NULL,
-    "last_id" INT = NULL
+    "last_id" BIGINT = NULL
 )
 RETURNS nfttracker_endpoints.trx_result[] 
 -- openapi-generated-code-end
